Migrate device_port_forwarder to null safety (#78949)
