+++ packages/flutter_tools/lib/src/protocol_discovery.dart
@@ -11,6 +11,7 @@ import 'package:meta/meta.dart';
 import 'base/io.dart';
 import 'base/logger.dart';
 import 'device.dart';
+import 'device_port_forwader.dart';
 import 'globals.dart' as globals;
 
 /// Discovers a specific service protocol on a device, and forwards the service
