+++ packages/flutter_tools/lib/src/web/web_device.dart
@@ -16,6 +16,7 @@ import '../base/platform.dart';
 import '../base/version.dart';
 import '../build_info.dart';
 import '../device.dart';
+import '../device_port_forwader.dart';
 import '../features.dart';
 import '../project.dart';
 import 'chrome.dart';
