+++ packages/flutter_tools/test/general.shard/android/android_device_port_forwarder_test.dart
@@ -7,7 +7,7 @@
 import 'package:flutter_tools/src/android/android_device.dart';
 import 'package:flutter_tools/src/base/io.dart';
 import 'package:flutter_tools/src/base/logger.dart';
-import 'package:flutter_tools/src/device.dart';
+import 'package:flutter_tools/src/device_port_forwader.dart';
 
 import '../../src/common.dart';
 import '../../src/context.dart';
